Fix panic in sign_es if r or s are too short
When building the ECSDA signature, in some cases 'r' and/or 's' may be
less than 32 bytes long.  In that case they need to be padded to 32
bytes to ensure the full signature is 64 bytes long.

The existing code copes with that by taking the length into account when
splicing the values into the signature, however it starts with an empty
vector, so if 's' is less than 32 bytes long then the splice starts
after the end of the signature and triggers a panic.  A similar issue
affects the splice of 'r'.

Simply initializing the signature to be 64 bytes of 0s isn't sufficient
as when splicing 's' it includes a range up to the end of the signature.
This fix therefore also changes the splice ranges to only affect the
relevant section of the signature.